requirements/base.txt
Django==3.2.9
django-environ==0.8.1
django-filter==21.1
requests==2.26.0

# DRF
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1

# psycopg2-binary==2.9.1